# Accept multiple aggregations within a single query