… aggregate_test.go
… query/logicalplan/builder_test.go
… query/logicalplan/logicalplan_test.go
… query/logicalplan/optimize_test.go
… query/logicalplan/validate_test.go