FrostDB should be able to accept queries like:
```
select sum(value), count(value), max(value) group by labels.label2
select sum(value) as value_sum, count(value) as value_count group by second(3)
```

For this, to work FrostDB needs to accept multiple aggregations within a single query. 
FrostDBs Aggregate interface has to be updated:
```diff
-Aggregate(aggExpr logicalplan.Expr, groupExprs ...logicalplan.Expr) Builder
+Aggregate(aggExpr, groupExprs []logicalplan.Expr) Builder
```

To make our lives easier to figure out all the things that belong to a single column that's being aggregated, I've introduced `AggregateColumn`. 

Benchmarks show no significant impact:
```
name                  old time/op    new time/op    delta
Aggregation/sum-24      1.84ms ± 3%    1.87ms ± 1%    ~     (p=0.065 n=10+9)
Aggregation/count-24    1.85ms ± 1%    1.85ms ± 1%    ~     (p=0.489 n=9+9)
Aggregation/max-24      1.83ms ± 2%    1.86ms ± 1%  +1.25%  (p=0.004 n=10+9)

name                  old alloc/op   new alloc/op   delta
Aggregation/sum-24      1.45MB ± 0%    1.46MB ± 0%  +0.16%  (p=0.000 n=10+10)
Aggregation/count-24    1.45MB ± 0%    1.46MB ± 0%  +0.17%  (p=0.000 n=10+10)
Aggregation/max-24      1.45MB ± 0%    1.46MB ± 0%  +0.17%  (p=0.000 n=10+10)

name                  old allocs/op  new allocs/op  delta
Aggregation/sum-24       1.43k ± 0%     1.50k ± 0%  +5.23%  (p=0.000 n=10+9)
Aggregation/count-24     1.43k ± 0%     1.50k ± 0%  +5.25%  (p=0.000 n=10+10)
Aggregation/max-24       1.43k ± 0%     1.50k ± 0%  +5.25%  (p=0.000 n=9+8)
```