@metalmatze 2022-12-12
There are some TODOs left in the validation we should fix. I think before merging this PR, otherwise, it won't happen :grimacing: